Update customer-table.component.ts
Customer list gets reloaded once dialog is closed.
Previously, when viewing specific customer details, if any changes were made in the input fields, these changes  would be visible on the table after dialog is closed until the page is refreshed.
Now the list is reloaded after the dialog is closed so that the table would reflect correct information.